cmd/satellite: organize subcommands
