+++ cmd/satellite/main.go
@@ -23,6 +23,7 @@ import (
 	"storj.io/private/cfgstruct"
 	"storj.io/private/process"
 	"storj.io/private/version"
+	"storj.io/storj/cmd/satellite/billing"
 	"storj.io/storj/cmd/satellite/reports"
 	"storj.io/storj/pkg/cache"
 	"storj.io/storj/pkg/revocation"
@@ -32,9 +33,7 @@ import (
 	"storj.io/storj/satellite/compensation"
 	"storj.io/storj/satellite/metainfo"
 	"storj.io/storj/satellite/orders"
-	"storj.io/storj/satellite/payments/stripecoinpayments"
 	"storj.io/storj/satellite/satellitedb"
-	"storj.io/storj/satellite/satellitedb/dbx"
 )
 
 // Satellite defines satellite configuration
@@ -535,12 +534,6 @@ func cmdNodeUsage(cmd *cobra.Command, args []string) (err error) {
 	return generateNodeUsageCSV(ctx, start, end, file)
 }
 
-func cmdStripeCustomer(cmd *cobra.Command, args []string) (err error) {
-	ctx, _ := process.Ctx(cmd)
-
-	return generateStripeCustomers(ctx)
-}
-
 func cmdGenerateInvoices(cmd *cobra.Command, args []string) (err error) {
 	ctx, _ := process.Ctx(cmd)
 
@@ -622,69 +615,40 @@ func cmdValueAttribution(cmd *cobra.Command, args []string) (err error) {
 	return reports.GenerateAttributionCSV(ctx, partnerAttribtionCfg.Database, partnerID, start, end, file)
 }
 
+func cmdStripeCustomer(cmd *cobra.Command, args []string) (err error) {
+	ctx, _ := process.Ctx(cmd)
+
+	return billing.GenerateStripeCustomers(ctx, runCfg.Database, runCfg.Payments)
+}
+
 func cmdPrepareCustomerInvoiceRecords(cmd *cobra.Command, args []string) (err error) {
 	ctx, _ := process.Ctx(cmd)
 
-	period, err := parseBillingPeriod(args[0])
-	if err != nil {
-		return errs.New("invalid period specified: %v", err)
-	}
-
-	return runBillingCmd(func(payments *stripecoinpayments.Service, _ *dbx.DB) error {
-		return payments.PrepareInvoiceProjectRecords(ctx, period)
-	})
+	return billing.PrepareCustomerInvoiceItems(ctx, args[0], runCfg.Database, runCfg.Payments)
 }
 
 func cmdCreateCustomerInvoiceItems(cmd *cobra.Command, args []string) (err error) {
 	ctx, _ := process.Ctx(cmd)
 
-	period, err := parseBillingPeriod(args[0])
-	if err != nil {
-		return errs.New("invalid period specified: %v", err)
-	}
-
-	return runBillingCmd(func(payments *stripecoinpayments.Service, _ *dbx.DB) error {
-		return payments.InvoiceApplyProjectRecords(ctx, period)
-	})
+	return billing.CreateCustomerInvoiceItems(ctx, args[0], runCfg.Database, runCfg.Payments)
 }
 
 func cmdCreateCustomerInvoiceCoupons(cmd *cobra.Command, args []string) (err error) {
 	ctx, _ := process.Ctx(cmd)
 
-	period, err := parseBillingPeriod(args[0])
-	if err != nil {
-		return errs.New("invalid period specified: %v", err)
-	}
-
-	return runBillingCmd(func(payments *stripecoinpayments.Service, _ *dbx.DB) error {
-		return payments.InvoiceApplyCoupons(ctx, period)
-	})
+	return billing.CreateCustomerInvoiceCoupons(ctx, args[0], runCfg.Database, runCfg.Payments)
 }
 
 func cmdCreateCustomerInvoiceCredits(cmd *cobra.Command, args []string) (err error) {
 	ctx, _ := process.Ctx(cmd)
 
-	period, err := parseBillingPeriod(args[0])
-	if err != nil {
-		return errs.New("invalid period specified: %v", err)
-	}
-
-	return runBillingCmd(func(payments *stripecoinpayments.Service, _ *dbx.DB) error {
-		return payments.InvoiceApplyCredits(ctx, period)
-	})
+	return billing.CreateCustomerInvoiceCredits(ctx, args[0], runCfg.Database, runCfg.Payments)
 }
 
 func cmdCreateCustomerInvoices(cmd *cobra.Command, args []string) (err error) {
 	ctx, _ := process.Ctx(cmd)
 
-	period, err := parseBillingPeriod(args[0])
-	if err != nil {
-		return errs.New("invalid period specified: %v", err)
-	}
-
-	return runBillingCmd(func(payments *stripecoinpayments.Service, _ *dbx.DB) error {
-		return payments.CreateInvoices(ctx, period)
-	})
+	return billing.CreateCustomerInvoices(ctx, args[0], runCfg.Database, runCfg.Payments)
 }
 
 func main() {
